+++ priv/std_lib/exceptions.ex
@@ -1,0 +1,7 @@
+defmodule ElixirScript.ArgumentError do
+  defexception message: "argument error"
+end
+
+defmodule ElixirScript.RuntimeError do
+  defexception message: "runtime error"
+end