+++ priv/std_lib/range.ex
@@ -3,10 +3,10 @@ defmodule ElixirScript.Range do
   defstruct first: nil, last: nil
 
   def new(first, last) do
-    %ElixirScript.Range{first: first, last: last}
+    %Range{first: first, last: last}
   end
 
-  def range?(%ElixirScript.Range{}), do: true
+  def range?(%Range{}), do: true
   def range?(_), do: false
 
 end