+++ src/javascript/lib/core.js
@@ -5,10 +5,24 @@ import SpecialForms from './core/special_forms';
 
 class Integer {}
 class Float {}
-const global = Functions.get_global();
 
-global.__elixirscript_store__ = new Map();
-global.__elixirscript_names__ = new Map();
+function get_global() {
+  if (typeof self !== 'undefined') {
+    return self;
+  } else if (typeof window !== 'undefined') {
+    return window;
+  } else if (typeof global !== 'undefined') {
+    return global;
+  }
+
+  console.warn('No global state found');
+  return null;
+}
+
+const globalState = get_global();
+
+globalState.__elixirscript_store__ = new Map();
+globalState.__elixirscript_names__ = new Map();
 
 export default {
   Tuple: ErlangTypes.Tuple,
@@ -19,5 +33,5 @@ export default {
   Float,
   Functions,
   SpecialForms,
-  global,
+  global: globalState,
 };