+++ test/translator/try_test.exs
@@ -13,11 +13,21 @@ defmodule ElixirScript.Translator.Try.Test do
     end
 
     js_code = """
+<<<<<<< HEAD
     Bootstrap.Core.SpecialForms._try(async function() {
         return 1;
     }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.type(ArgumentError, {})], async function() {
     return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
     })), null, null, null)
+=======
+         Bootstrap.Core.SpecialForms._try(function() {
+             return 1;
+         }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([{
+             [Symbol.for('__struct__')]: Symbol.for('Elixir.ElixirScript.ArgumentError')
+         }], function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
+         })), null, null, null)
+>>>>>>> master
     """
 
     assert_translation(ex_ast, js_code)
@@ -35,11 +45,21 @@ defmodule ElixirScript.Translator.Try.Test do
     end
 
     js_code = """
+<<<<<<< HEAD
     Bootstrap.Core.SpecialForms._try(async function() {
         return 1;
     }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.type(ArgumentError, {})], async function() {
     return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
     })), null, null, null)
+=======
+         Bootstrap.Core.SpecialForms._try(function() {
+             return 1;
+         }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([{
+             [Symbol.for('__struct__')]: Symbol.for('Elixir.ElixirScript.ArgumentError')
+         }], function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
+         })), null, null, null)
+>>>>>>> master
     """
 
     assert_translation(ex_ast, js_code)
@@ -61,8 +81,13 @@ defmodule ElixirScript.Translator.Try.Test do
              return     1;
            },Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.variable()],async function(x)    {
            return     Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
+<<<<<<< HEAD
            },async function(x)    {
            return     Elixir.ElixirScript.Bootstrap.Functions.__load(Elixir).contains(x,Object.freeze([ArgumentError.create(Object.freeze({}))]));
+=======
+           },function(x)    {
+           return Elixir.ElixirScript.Bootstrap.Functions.__load(Elixir).contains(x, Object.freeze([Elixir.ElixirScript.ArgumentError.__load(Elixir).__struct__(Object.freeze({}))]));
+>>>>>>> master
            })),null,null,null)
     """
 
@@ -106,13 +131,15 @@ defmodule ElixirScript.Translator.Try.Test do
     end
 
     js_code = """
-    Bootstrap.Core.SpecialForms._try(async function() {
-        return 1;
-    }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.type(ArgumentError, {})], async function() {
-    return Elixir.ElixirScript.IO.__load(Elixir).puts('ArgumentError');
-    }), Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.variable()], async function(x) {
-    return Elixir.ElixirScript.IO.__load(Elixir).puts('x');
-    })), null, null, null)
+         Bootstrap.Core.SpecialForms._try(function() {
+             return 1;
+         }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([{
+             [Symbol.for('__struct__')]: Symbol.for('Elixir.ElixirScript.ArgumentError')
+         }], function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('ArgumentError');
+         }), Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.variable()], function(x) {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('x');
+         })), null, null, null)
     """
 
     assert_translation(ex_ast, js_code)
@@ -132,13 +159,15 @@ defmodule ElixirScript.Translator.Try.Test do
     end
 
     js_code = """
-    Bootstrap.Core.SpecialForms._try(async function() {
-        return 1;
-    }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.type(ArgumentError, {})], async function() {
-    return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
-    })), null, null, async function() {
-    return Elixir.ElixirScript.IO.__load(Elixir).puts('This is printed regardless if it failed or succeed');
-    })
+         Bootstrap.Core.SpecialForms._try(function() {
+             return 1;
+         }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([{
+             [Symbol.for('__struct__')]: Symbol.for('Elixir.ElixirScript.ArgumentError')
+         }], function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
+         })), null, null, function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('This is printed regardless if it failed or succeed');
+         })
     """
 
     assert_translation(ex_ast, js_code)
@@ -207,13 +236,15 @@ defmodule ElixirScript.Translator.Try.Test do
     end
 
     js_code = """
-    Bootstrap.Core.SpecialForms._try(async function() {
-        return 1;
-    }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Bootstrap.Core.Patterns.type(ArgumentError, {})], async function() {
-    return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
-    })), Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Symbol.for('throw'), Symbol.for('Error')], async function() {
-    return Elixir.ElixirScript.IO.__load(Elixir).puts('caught error');
-    })), null, null)
+         Bootstrap.Core.SpecialForms._try(function() {
+             return 1;
+         }, Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([{
+             [Symbol.for('__struct__')]: Symbol.for('Elixir.ElixirScript.ArgumentError')
+         }], function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('Invalid argument given');
+         })), Bootstrap.Core.Patterns.defmatch(Bootstrap.Core.Patterns.clause([Symbol.for('throw'), Symbol.for('Error')], function() {
+             return Elixir.ElixirScript.IO.__load(Elixir).puts('caught error');
+         })), null, null)
     """
 
     assert_translation(ex_ast, js_code)